factor out interface
