+++ core/src/main/java/io/opentelemetry/android/OpenTelemetryRumBuilder.java
@@ -31,6 +31,7 @@ import io.opentelemetry.android.internal.services.ServiceManager;
 import io.opentelemetry.android.internal.services.ServiceManagerImpl;
 import io.opentelemetry.android.internal.services.periodicwork.PeriodicWorkService;
 import io.opentelemetry.android.session.SessionManager;
+import io.opentelemetry.android.session.SessionManagerImpl;
 import io.opentelemetry.android.session.SessionProvider;
 import io.opentelemetry.api.baggage.propagation.W3CBaggagePropagator;
 import io.opentelemetry.api.trace.propagation.W3CTraceContextPropagator;
@@ -307,7 +308,7 @@ public final class OpenTelemetryRumBuilder {
         initializationEvents.spanExporterInitialized(spanExporter);
 
         SessionManager sessionManager =
-                SessionManager.create(timeoutHandler, config.getSessionTimeout().toNanos());
+                SessionManagerImpl.create(timeoutHandler, config.getSessionTimeout().toNanos());
 
         OpenTelemetrySdk sdk =
                 OpenTelemetrySdk.builder()
